Update Dockerfile and package.json dependencies
This commit updates the Dockerfile to use a newer version of the frontend-builder. It also updates the package.json file to use the latest versions of various dependencies. The changes include both minor and patch updates. The use of caret (^) in the versioning ensures that we use the most recent major version. This should improve the overall performance and security of the application.